requirements.txt
fastapi
uvicorn
google-generativeai
sqlalchemy
pg8000
cloud-sql-python-connector
python-dotenv
pydantic